requirements.txt
kafka-python==1.4.7
psutil==5.6.7
psycopg2==2.8.4
ipython